+++ src/glview.h
@@ -283,6 +283,7 @@ public:
 		float rotSpd = 45;
 
 		UpAxis upAxis = ZAxis;
+		int	startupDirection = 1;	// +Y
 
 		static float	vertexPointSize;
 		static float	tbnPointSize;
@@ -294,6 +295,8 @@ public:
 		static float	lineWidthGrid1;
 		static float	lineWidthGrid2;
 		static float	lineWidthSelect;
+		static float	zoomInScale;
+		static float	zoomOutScale;
 	} cfg;
 
 private slots: